requirements.txt
numpy
matplotlib
seaborn
plotly
streamlit
scikit-learn
tensorflow-cpu
keras
protobuf
pandas
pandas-profiling
ppscore
feature-engine
imbalanced-learn
xgboost
yellowbrick
Jinja2
MarkupSafe
ipywidgets
